angularified + video directive done
Sorted out dir structure (again) as part of a refit to use angular. Restructured JS files to use angular as an npm module along with browserify for requiring it.

Also did a lot of work on the video directive. Only takes one src right now so need to expand it to take several, but currently works, with mute button and black striped overlay.