feat: adds email validation and activation
When registering the email address is now checked with a confirm email. This then sends email with UUID to address provided. Unique URL will then activate the user within the DB. User is created with a TTL. Following the link in the email marks activated field as true and removes the TTL field.